rename madplot
